==> 4. data discritization.ipynb <==
Input file: road profile data from(2)
output file: discrete road profile data to()

import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns
import numpy as np

output_dir = 'D:/Thesis/files_output_dir/output_preprocessing/'
# Load the history workday files into DataFrames
level1_road_history_workday = pd.read_csv(output_dir + 'level1_road_history_workday.csv') 
level2_road_history_workday = pd.read_csv(output_dir + 'level2_road_history_workday.csv')
level3_road_history_workday = pd.read_csv(output_dir + 'level3_road_history_workday.csv')

output_dir = 'D:/Thesis/files_output_dir/output_preprocessing/'
# Load the history holiday files into DataFrames
level1_road_history_holiday = pd.read_csv(output_dir + 'level1_road_history_holiday.csv') 
level2_road_history_holiday = pd.read_csv(output_dir + 'level2_road_history_holiday.csv')
level3_road_history_holiday = pd.read_csv(output_dir + 'level3_road_history_holiday.csv')

## Discritization Fuction

# Functions for data discretization

def discreteLength(length):
    if length < 100:
        return 0
    else:
        return 1

def discreteLane(lane):
    if lane < 2:
        return 0
    if lane < 3:
        return 1
    if lane < 4:
        return 2
    else:
        return 3

def discreteLevel(level):
    if level == 0 or level == 1:
        return 1
    if level == 2:
        return 2
    return 3

def discreteTor(tor):
    if tor < 1.25:
        return 0
    else:
        return 1
    
def discreteConnections(cnum):
    if cnum < 4:
        return 0
    if cnum < 5:
        return 1
    else:
        return 2

def discreteSumPOI(num):
    if num < 5:
        return 0
    return 1

def discretePOI(num):
    if num < 2:
        return 0
    if num < 4:
        return 1
    return 2

def discreteTime(timeSlot):
    if timeSlot < 7 * 6: #hour * hourly time slot num.
        return 0
    if timeSlot < 10 *6:
        return 1
    if timeSlot < 16 * 6:
        return 2
    if timeSlot < 20 * 6:
        return 3
    return 4

def discreteSup(support, level):
    if level == 1:
        if support < 3:
            return 0
        if support < 6:
            return 1
        if support < 10:
            return 2
        if support < 15:
            return 3
        if support < 25:
            return 4
        if support < 50:
            return 5
        return 6
    else:
        if level == 2:
            if support < 3:
                return 0
            if support < 6:
                return 1
            if support < 9:
                return 2
            if support < 12:
                return 3
            if support < 20:
                return 4
            if support < 40:
                return 5
            return 6
        else:
            if support < 1:
                return 0
            if support < 2:
                return 1
            if support < 3:
                return 2
            if support < 5:
                return 3
            if support < 10:
                return 4
            if support < 20:
                return 5
            return 6

# discretize the average speed
def discreteSpeed(speed, level):
    if level == 0 or level == 1:
        if speed < 10:
            return 0
        if speed < 15:
            return 1
        if speed < 17:
            return 2
        if speed < 19:
            return 3
        if speed < 21:
            return 4
        return 5
    else:
        if level == 2:
            if speed < 4:
                return 0
            if speed < 6:
                return 1
            if speed < 8:
                return 2
            if speed < 10:
                return 3
            if speed < 12:
                return 4
            return 5
        else:
            if speed < 2:
                return 0
            if speed < 3.5:
                return 1
            if speed < 5:
                return 2
            if speed < 6.5:
                return 3
            if speed < 8:
                return 4
            return 5
def discreteMaxSpeed(speed, level):
    if level == 0 or level == 1:
        if speed <= 5:
            return 0
        if speed < 10:
            return 1
        if speed <= 15:
            return 2
        if speed < 20:
            return 3
        if speed < 25:
            return 4
        return 5
    else:
        if level == 2:
            if speed < 4:
                return 0
            if speed < 6:
                return 1
            if speed < 8:
                return 2
            if speed < 10:
                return 3
            if speed < 12:
                return 4
            return 5
        else:
            if speed < 2:
                return 0
            if speed < 3.5:
                return 1
            if speed < 5:
                return 2
            if speed < 6.5:
                return 3
            if speed < 8:
                return 4
            return 5
def discreteStdSpeed(speed, level):
    if level == 0 or level == 1:
        if speed < 2:
             return 0
        if speed < 3.5:
            return 1
        if speed < 5:
            return 2
        if speed < 6.5:
            return 3
        if speed < 8:
            return 4
        return 5
    else:
        if level == 2:
            if speed < 4:
                return 0
            if speed < 6:
                return 1
            if speed < 8:
                return 2
            if speed < 10:
                return 3
            if speed < 12:
                return 4
            return 5
        else:
            if speed < 2:
                return 0
            if speed < 3.5:
                return 1
            if speed < 5:
                return 2
            if speed < 6.5:
                return 3
            if speed < 8:
                return 4
            return 5

# This one for training data only
def discreteVolume(vol, level):
    if level == 1:
        if vol < 15:
            return 0
        if vol < 18.67:
            return 1
        if vol < 21.82:
            return 2
        if vol < 25.48:
            return 3
        return 4
    else:
        if level == 2:
            if vol < 3.03:
                return 0
            if vol < 6.19:
                return 1
            if vol < 8.91:
                return 2
            if vol < 12.07:
                return 3
            return 4
        else:
            if vol < 1.03:
                return 0
            if vol < 2.89:
                return 1
            if vol < 4.49:
                return 2
            if vol < 6.35:
                return 3
            return 4

# Optimized Functions for data discretization

def discreteLength(length):
    return 1 if length < 100 else 2

def discreteLane(lane):
    return min(lane, 3)  # Lane values are capped at 3

def discreteLevel(level):
    return min(level, 3)  # Levels 1 -> 1; 2 -> 2; 3 and above -> 3

def discreteDirrection(dir):
    return (dir+1)

def discreteTor(tor):
    return 1 if tor < 1.25 else 2

def discreteConnections(cnum):
    return 1 if cnum < 4 else (2 if cnum < 5 else 3)

def discreteSumPOI(num):
    return 1 if num < 5 else 2

def discretePOI(num):
    if num < 2:
        return 1
    elif num < 4:
        return 2
    return 3

def discretize_weighted_POI(weighted_POI, level):
    if level == 1:
        # Level 1: Custom bins
        bins = [0, 1, 5, 10, 20, np.inf]  # 0 -> 1, 1-5 -> 2, 5-10 -> 3, 10-20 -> 4, >20 -> 5
    elif level == 2:
        # Level 2: Different bins based on data distribution
        bins = [0, 2, 8, 15, 25, np.inf]  # Adjusted bins for level 2
    elif level >= 3:
        # Level 3: Another set of bins
        bins = [0, 3, 10, 18, 30, np.inf]  # Adjusted bins for level 3
    
    # Use numpy's digitize to convert continuous values to discrete bins
    discretized_values = np.digitize(weighted_POI, bins)
    
    return discretized_values

def discreteTime(timeSlot):
    if timeSlot < 7 * 6:
        return 1
    elif timeSlot < 10 * 6:
        return 2
    elif timeSlot < 16 * 6:
        return 3
    elif timeSlot < 20 * 6:
        return 4
    return 5

def discreteSup(support, level):
    bins = [
        [3, 6, 10, 15, 25, 50],  # For level 1
        [3, 6, 9, 12, 20, 40],   # For level 2
        [1, 2, 3, 5, 10, 20]     # For level 3+
    ]
    thresholds = bins[min(level - 1, 2)]
    for i, threshold in enumerate(thresholds):
        if support < threshold:
            return i + 1
    return len(thresholds) + 1

def discreteSpeed(speed, level):
    bins = [
        [10, 15, 17, 19, 21],  # For level 0, 1
        [4, 6, 8, 10, 12],     # For level 2
        [2, 3.5, 5, 6.5, 8]    # For level 3+
    ]
    thresholds = bins[min(level-1, 2)]
    for i, threshold in enumerate(thresholds):
        if speed < threshold:
            return i + 1
    return len(thresholds) + 1

def discreteMaxSpeed(speed, level):
    if level in [0, 1]:
        bins = [5, 10, 15, 20, 25]
    elif level == 2:
        bins = [4, 6, 8, 10, 12]
    elif level >= 3:
        bins = [2, 3.5, 5, 6.5, 8]
    
    for i, threshold in enumerate(bins):
        if speed <= threshold:
            return i + 1
    return len(bins) + 1

def discreteStdSpeed(speed, level):
    bins = [
        [2, 3.5, 5, 6.5, 8],  # For level 0, 1
        [4, 6, 8, 10, 12],    # For level 2
        [2, 3.5, 5, 6.5, 8]   # For level 3+
    ]
    thresholds = bins[min(level-1, 2)]
    for i, threshold in enumerate(thresholds):
        if speed < threshold:
            return i + 1
    return len(thresholds) + 1


# **high_way**

level1_road_history_workday.columns

'''
['road_id', 'length', 'lanes', 'max_speed', 'dir', 'level', 'tor',
       'n_connnections', 'school', 'ofc', 'bnk', 'mall', 'rest', 'gas_st',
       'spot', 'hotel', 'trans', 'living_ser', 'sum_poi', 'weighted_POI',
       'time', 'avg_speed', 'avg_speed std', 'no_taxi_car'],
      dtype='object')
'''

## Workday Level 1

# Apply discretization to the level1_road_history_workday DataFrame
level=1
level1_road_history_workday['length'] = level1_road_history_workday['length'].apply(lambda x: discreteLength(x))
level1_road_history_workday['lanes'] = level1_road_history_workday['lanes'].apply(lambda x: discreteLane(x))
level1_road_history_workday['max_speed'] = level1_road_history_workday['max_speed'].apply(lambda x: discreteMaxSpeed(x, 1))
level1_road_history_workday['level'] = level1_road_history_workday['level'].apply(lambda x: discreteLevel(level))
level1_road_history_workday['dir'] = level1_road_history_workday['dir'].apply(lambda x: discreteDirrection(x))
level1_road_history_workday['tor'] = level1_road_history_workday['tor'].apply(lambda x: discreteTor(x))
level1_road_history_workday['n_connnections'] = level1_road_history_workday['n_connnections'].apply(lambda x: discreteConnections(x))
level1_road_history_workday['school'] = level1_road_history_workday['school'].apply(lambda x: discretePOI(x))
level1_road_history_workday['ofc'] = level1_road_history_workday['ofc'].apply(lambda x: discretePOI(x))
level1_road_history_workday['bnk'] = level1_road_history_workday['bnk'].apply(lambda x: discretePOI(x))
level1_road_history_workday['mall'] = level1_road_history_workday['mall'].apply(lambda x: discretePOI(x))
level1_road_history_workday['rest'] = level1_road_history_workday['rest'].apply(lambda x: discretePOI(x))
level1_road_history_workday['gas_st'] = level1_road_history_workday['gas_st'].apply(lambda x: discretePOI(x))
level1_road_history_workday['spot'] = level1_road_history_workday['spot'].apply(lambda x: discretePOI(x))
level1_road_history_workday['hotel'] = level1_road_history_workday['hotel'].apply(lambda x: discretePOI(x))
level1_road_history_workday['trans'] = level1_road_history_workday['trans'].apply(lambda x: discretePOI(x))
level1_road_history_workday['living_ser'] = level1_road_history_workday['living_ser'].apply(lambda x: discretePOI(x))
level1_road_history_workday['sum_poi'] = level1_road_history_workday['sum_poi'].apply(lambda x: discreteSumPOI(x))
level1_road_history_workday['weighted_POI']= level1_road_history_workday['weighted_POI'].apply(lambda x: discretize_weighted_POI(x, level))
level1_road_history_workday['time'] = level1_road_history_workday['time'].apply(lambda x: discreteTime(x))
level1_road_history_workday['avg_speed'] = level1_road_history_workday['avg_speed'].apply(lambda x: discreteSpeed(x, level))
level1_road_history_workday['avg_speed std'] = level1_road_history_workday['avg_speed std'].apply(lambda x: discreteStdSpeed(x, level))
level1_road_history_workday['no_taxi_car'] = level1_road_history_workday['no_taxi_car'].apply(lambda x: discreteSup(x, level))

## holiday Level-1

level=1

column_function_mapping = {
    'length': discreteLength,
    'lanes': discreteLane,
    'max_speed': lambda x: discreteMaxSpeed(x, level),
    'level': lambda x: discreteLevel(level),
    'dir': discreteDirrection,
    'tor': discreteTor,
    'n_connnections': discreteConnections,
    'school': discretePOI,
    'ofc': discretePOI,
    'bnk': discretePOI,
    'mall': discretePOI,
    'rest': discretePOI,
    'gas_st': discretePOI,
    'spot': discretePOI,
    'hotel': discretePOI,
    'trans': discretePOI,
    'living_ser': discretePOI,
    'sum_poi': discreteSumPOI,
    'weighted_POI': lambda x: discretize_weighted_POI(x, level),
    'time': discreteTime,
    'avg_speed': lambda x: discreteSpeed(x, level),
    'avg_speed std': lambda x: discreteStdSpeed(x, level),
    'no_taxi_car': lambda x: discreteSup(x, level)
}

# Iterate over the mapping and apply the functions to each column
for column, func in column_function_mapping.items():
    level1_road_history_holiday[column] = level1_road_history_holiday[column].apply(func)


level1_road_history_holiday.head()

# **major-road**

## Workday Level 2

# Apply discretization to the level2_road_history_workday DataFrame
level=2
level2_road_history_workday['length'] = level2_road_history_workday['length'].apply(lambda x: discreteLength(x))
level2_road_history_workday['lanes'] = level2_road_history_workday['lanes'].apply(lambda x: discreteLane(x))
level2_road_history_workday['max_speed'] = level2_road_history_workday['max_speed'].apply(lambda x: discreteMaxSpeed(x, level))
level2_road_history_workday['level'] = level2_road_history_workday['level'].apply(lambda x: discreteLevel(level))
level2_road_history_workday['dir'] = level2_road_history_workday['dir'].apply(lambda x: discreteDirrection(x))
level2_road_history_workday['tor'] = level2_road_history_workday['tor'].apply(lambda x: discreteTor(x))
level2_road_history_workday['n_connnections'] = level2_road_history_workday['n_connnections'].apply(lambda x: discreteConnections(x))
level2_road_history_workday['school'] = level2_road_history_workday['school'].apply(lambda x: discretePOI(x))
level2_road_history_workday['ofc'] = level2_road_history_workday['ofc'].apply(lambda x: discretePOI(x))
level2_road_history_workday['bnk'] = level2_road_history_workday['bnk'].apply(lambda x: discretePOI(x))
level2_road_history_workday['mall'] = level2_road_history_workday['mall'].apply(lambda x: discretePOI(x))
level2_road_history_workday['rest'] = level2_road_history_workday['rest'].apply(lambda x: discretePOI(x))
level2_road_history_workday['gas_st'] = level2_road_history_workday['gas_st'].apply(lambda x: discretePOI(x))
level2_road_history_workday['spot'] = level2_road_history_workday['spot'].apply(lambda x: discretePOI(x))
level2_road_history_workday['hotel'] = level2_road_history_workday['hotel'].apply(lambda x: discretePOI(x))
level2_road_history_workday['trans'] = level2_road_history_workday['trans'].apply(lambda x: discretePOI(x))
level2_road_history_workday['living_ser'] = level2_road_history_workday['living_ser'].apply(lambda x: discretePOI(x))
level2_road_history_workday['sum_poi'] = level2_road_history_workday['sum_poi'].apply(lambda x: discreteSumPOI(x))
level2_road_history_workday['weighted_POI'] = level2_road_history_workday['weighted_POI'].apply(lambda x: discretize_weighted_POI(x, level))
level2_road_history_workday['time'] = level2_road_history_workday['time'].apply(lambda x: discreteTime(x))
level2_road_history_workday['avg_speed'] = level2_road_history_workday['avg_speed'].apply(lambda x: discreteSpeed(x, level))
level2_road_history_workday['avg_speed std'] = level2_road_history_workday['avg_speed std'].apply(lambda x: discreteStdSpeed(x, level))
level2_road_history_workday['no_taxi_car'] = level2_road_history_workday['no_taxi_car'].apply(lambda x: discreteSup(x, level))

level2_road_history_workday.head()

## Holiday Level-2

level=2

column_function_mapping = {
    'length': discreteLength,
    'lanes': discreteLane,
    'max_speed': lambda x: discreteMaxSpeed(x, level),
    'level': lambda x: discreteLevel(level),
    'dir': discreteDirrection,
    'tor': discreteTor,
    'n_connnections': discreteConnections,
    'school': discretePOI,
    'ofc': discretePOI,
    'bnk': discretePOI,
    'mall': discretePOI,
    'rest': discretePOI,
    'gas_st': discretePOI,
    'spot': discretePOI,
    'hotel': discretePOI,
    'trans': discretePOI,
    'living_ser': discretePOI,
    'sum_poi': discreteSumPOI,
    'weighted_POI': lambda x: discretize_weighted_POI(x, level),
    'time': discreteTime,
    'avg_speed': lambda x: discreteSpeed(x, level),
    'avg_speed std': lambda x: discreteStdSpeed(x, level),
    'no_taxi_car': lambda x: discreteSup(x, level)
}

# Iterate over the mapping and apply the functions to each column
for column, func in column_function_mapping.items():
    level2_road_history_holiday[column] = level2_road_history_holiday[column].apply(func)


level2_road_history_holiday.head()

# **small-road**

level = 3

level3_road_history_workday['length'] = level3_road_history_workday['length'].apply(lambda x: discreteLength(x))
level3_road_history_workday['lanes'] = level3_road_history_workday['lanes'].apply(lambda x: discreteLane(x))
level3_road_history_workday['max_speed'] = level3_road_history_workday['max_speed'].apply(lambda x: discreteMaxSpeed(x, level))
level3_road_history_workday['level'] = level3_road_history_workday['level'].apply(lambda x: discreteLevel(level))
level3_road_history_workday['dir'] = level3_road_history_workday['dir'].apply(lambda x: discreteDirrection(x))
level3_road_history_workday['tor'] = level3_road_history_workday['tor'].apply(lambda x: discreteTor(x))
level3_road_history_workday['n_connnections'] = level3_road_history_workday['n_connnections'].apply(lambda x: discreteConnections(x))
level3_road_history_workday['school'] = level3_road_history_workday['school'].apply(lambda x: discretePOI(x))
level3_road_history_workday['ofc'] = level3_road_history_workday['ofc'].apply(lambda x: discretePOI(x))
level3_road_history_workday['bnk'] = level3_road_history_workday['bnk'].apply(lambda x: discretePOI(x))
level3_road_history_workday['mall'] = level3_road_history_workday['mall'].apply(lambda x: discretePOI(x))
level3_road_history_workday['rest'] = level3_road_history_workday['rest'].apply(lambda x: discretePOI(x))
level3_road_history_workday['gas_st'] = level3_road_history_workday['gas_st'].apply(lambda x: discretePOI(x))
level3_road_history_workday['spot'] = level3_road_history_workday['spot'].apply(lambda x: discretePOI(x))
level3_road_history_workday['hotel'] = level3_road_history_workday['hotel'].apply(lambda x: discretePOI(x))
level3_road_history_workday['trans'] = level3_road_history_workday['trans'].apply(lambda x: discretePOI(x))
level3_road_history_workday['living_ser'] = level3_road_history_workday['living_ser'].apply(lambda x: discretePOI(x))
level3_road_history_workday['sum_poi'] = level3_road_history_workday['sum_poi'].apply(lambda x: discreteSumPOI(x))
level3_road_history_workday['weighted_POI'] = level3_road_history_workday['weighted_POI'].apply(lambda x: discretize_weighted_POI(x, level))
level3_road_history_workday['time'] = level3_road_history_workday['time'].apply(lambda x: discreteTime(x))
level3_road_history_workday['avg_speed'] = level3_road_history_workday['avg_speed'].apply(lambda x: discreteSpeed(x, level))
level3_road_history_workday['avg_speed std'] = level3_road_history_workday['avg_speed std'].apply(lambda x: discreteStdSpeed(x, level))
level3_road_history_workday['no_taxi_car'] = level3_road_history_workday['no_taxi_car'].apply(lambda x: discreteSup(x, level))

level3_road_history_workday.head()

## Holiday

level=3

column_function_mapping = {
    'length': discreteLength,
    'lanes': discreteLane,
    'max_speed': lambda x: discreteMaxSpeed(x, level),
    'level': lambda x: discreteLevel(level),
    'dir': discreteDirrection,
    'tor': discreteTor,
    'n_connnections': discreteConnections,
    'school': discretePOI,
    'ofc': discretePOI,
    'bnk': discretePOI,
    'mall': discretePOI,
    'rest': discretePOI,
    'gas_st': discretePOI,
    'spot': discretePOI,
    'hotel': discretePOI,
    'trans': discretePOI,
    'living_ser': discretePOI,
    'sum_poi': discreteSumPOI,
    'weighted_POI': lambda x: discretize_weighted_POI(x, level),
    'time': discreteTime,
    'avg_speed': lambda x: discreteSpeed(x, level),
    'avg_speed std': lambda x: discreteStdSpeed(x, level),
    'no_taxi_car': lambda x: discreteSup(x, level)
}

# Iterate over the mapping and apply the functions to each column
for column, func in column_function_mapping.items():
    level3_road_history_holiday[column] = level3_road_history_holiday[column].apply(func)

level3_road_history_holiday.head()

# Workday Data

output_dir = "D:/Thesis/files_output_dir/output_files/"
level1_road_history_workday.to_csv(output_dir +'discrete_level1_road_history_workday.csv', index=False)
level2_road_history_workday.to_csv(output_dir +'discrete_level2_road_history_workday.csv', index=False)
level3_road_history_workday.to_csv(output_dir +'discrete_level3_road_history_workday.csv', index=False)

# Holiday Data

output_dir = "D:/Thesis/files_output_dir/output_files/"
level1_road_history_holiday.to_csv(output_dir + 'discrete_level1_road_history_holiday.csv', index=False)
level2_road_history_holiday.to_csv(output_dir + 'discrete_level2_road_history_holiday.csv', index=False)
level3_road_history_holiday.to_csv(output_dir + 'discrete_level3_road_history_holiday.csv', index=False)
