==> 6. PGM volume inference.ipynb <==
Input: PGM model output--> road volume inference, road profile data
Output: infered road profile volume to (7)

import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns
import numpy as np
import scipy as sp
import pickle

file_path = "D:/Thesis/files_output_dir/output_files/"

output_dir = 'D:/Thesis/files_output_dir/output_preprocessing/'

level1_road_history_workday = pd.read_csv(output_dir + 'level1_road_history_workday.csv') 
level2_road_history_workday = pd.read_csv(output_dir + 'level2_road_history_workday.csv')
level3_road_history_workday = pd.read_csv(output_dir + 'level3_road_history_workday.csv')

# Load the history holiday files into DataFrames
level1_road_history_holiday = pd.read_csv(output_dir + 'level1_road_history_holiday.csv') 
level2_road_history_holiday = pd.read_csv(output_dir + 'level2_road_history_holiday.csv')
level3_road_history_holiday = pd.read_csv(output_dir + 'level3_road_history_holiday.csv')

level1_road_history_workday_inference = pd.read_csv(file_path + 'level1_road_history_workday_inference.csv')
print(level1_road_history_workday_inference['Volume'].value_counts(dropna=False))

level1_road_history_holiday_inference = pd.read_csv(file_path + 'level1_road_history_holiday_inference.csv')
print(level1_road_history_holiday_inference['Volume'].value_counts(dropna=False))

level2_road_history_workday_inference = pd.read_csv(file_path + 'level2_road_history_workday_inference.csv')
print(level2_road_history_workday_inference['Volume'].value_counts(dropna=False))

level2_road_history_holiday_inference = pd.read_csv(file_path + 'level2_road_history_holiday_inference.csv')
print(level2_road_history_holiday_inference['Volume'].value_counts(dropna=False))

level3_road_history_workday_inference = pd.read_csv(file_path + 'level3_road_history_workday_inference.csv')
print(level3_road_history_workday_inference['Volume'].value_counts(dropna=False))

level3_road_history_holiday_inference = pd.read_csv(file_path + 'level3_road_history_holiday_inference.csv')
print(level3_road_history_holiday_inference['Volume'].value_counts(dropna=False))

print(level2_road_history_holiday_inference.shape)
print(level2_road_history_holiday.shape)


level1_road_history_holiday_inference.head()
level1_road_history_holiday.head()

level1_road_history_workday['volume_claas'] = level1_road_history_workday_inference['Volume']
level1_road_history_workday['theta'] = level1_road_history_workday_inference['theta']

level1_road_history_holiday['volume_claas'] = level1_road_history_holiday_inference['Volume']
level1_road_history_holiday['theta'] = level1_road_history_holiday_inference['theta']

level1_road_history_holiday.head()

level2_road_history_workday['volume_claas'] = level2_road_history_workday_inference['Volume']
level2_road_history_workday['theta'] = level2_road_history_workday_inference['theta']

level2_road_history_holiday['volume_claas'] = level2_road_history_holiday_inference['Volume']
level2_road_history_holiday['theta'] = level2_road_history_holiday_inference['theta']

level3_road_history_workday['volume_claas'] = level3_road_history_workday_inference['Volume']
level3_road_history_workday['theta'] = level3_road_history_workday_inference['theta']

level3_road_history_holiday['volume_claas'] = level3_road_history_holiday_inference['Volume']
level3_road_history_holiday['theta'] = level3_road_history_holiday_inference['theta']

output_dir = 'D:/Thesis/files_output_dir/output_preprocessing/'
# Load from a file
with open(output_dir+ 'vol_distribution_data.pkl', 'rb') as file:
    data = pickle.load(file)

vol_dict_level1 = data['vol_dict1']
vol_dict_level2 = data['vol_dict2']
vol_dict_level3 = data['vol_dict3']
mean_level1 = data['mean1']
mean_level2 = data['mean2']
mean_level3 = data['mean3']


# Define the function with vol_dict as an additional parameter
def map_class_to_volume(cls, vol_dict):
    low, high = vol_dict[cls+1]
    return (low + high) / 2  # Use the midpoint of the range as the most probable value

# Apply the function using .map() and a lambda to pass vol_dict
level1_road_history_workday['volume'] = level1_road_history_workday['volume_claas'].map(lambda cls: map_class_to_volume(cls, vol_dict_level1))
level1_road_history_holiday['volume'] = level1_road_history_holiday['volume_claas'].map(lambda cls: map_class_to_volume(cls, vol_dict_level1))

level1_road_history_workday['volume'] = level1_road_history_workday['volume'] * level1_road_history_workday['lanes']
level1_road_history_holiday['volume'] = level1_road_history_holiday['volume'] * level1_road_history_holiday['lanes']

print(vol_dict_level1)
level1_road_history_workday.head()

# Apply the function using .map() and a lambda to pass vol_dict
level2_road_history_workday['volume'] = level2_road_history_workday['volume_claas'].map(lambda cls: map_class_to_volume(cls, vol_dict_level2))
# road volume calculation 
level2_road_history_workday['volume'] = level2_road_history_workday['volume'] * level2_road_history_workday['lanes']


# Apply the function using .map() and a lambda to pass vol_dict
level2_road_history_holiday['volume'] = level2_road_history_holiday['volume_claas'].map(lambda cls: map_class_to_volume(cls, vol_dict_level2))
# road volume calculation 
level2_road_history_holiday['volume'] = level2_road_history_holiday['volume'] * level2_road_history_holiday['lanes']

print(vol_dict_level2)
level2_road_history_holiday.head()

# Apply the function using .map() and a lambda to pass vol_dict
level3_road_history_workday['volume'] = level3_road_history_workday['volume_claas'].map(lambda cls: map_class_to_volume(cls, vol_dict_level3))
# road volume calculation 
level3_road_history_workday['volume'] = level3_road_history_workday['volume'] * level3_road_history_workday['lanes']

# Apply the function using .map() and a lambda to pass vol_dict
level3_road_history_holiday['volume'] = level3_road_history_holiday['volume_claas'].map(lambda cls: map_class_to_volume(cls, vol_dict_level3))
# road volume calculation 
level3_road_history_holiday['volume'] = level3_road_history_holiday['volume'] * level3_road_history_holiday['lanes']

print(vol_dict_level3)
level3_road_history_holiday.head()

# def get_volume_high(c):
#     vol = [18.7, 23.4, 25.75, 26.87, 29.7]
#     return vol[c-1];

# def get_volume_major(c):
#     vol = [2.2, 3.9, 5.95, 6.47, 7.2]
#     return vol[c-1];

# def get_volume_small(c):
#     vol = [0.77, 1.6, 3.1, 4.65, 5.7]
#     return vol[c-1];

level1_road_history_workday.to_csv(file_path + 'level1_road_history_workday_profile.csv', index=False)
level2_road_history_workday.to_csv(file_path + 'level2_road_history_workday_profile.csv', index=False)
level3_road_history_workday.to_csv(file_path + 'level3_road_history_workday_profile.csv', index=False)

level1_road_history_holiday.to_csv(file_path + 'level1_road_history_holiday_profile.csv', index=False)
level2_road_history_holiday.to_csv(file_path + 'level2_road_history_holiday_profile.csv', index=False)
level3_road_history_holiday.to_csv(file_path + 'level3_road_history_holiday_profile.csv', index=False)
