==> 8.3. road utilization view.ipynb <==
import pandas as pd

output_dir = 'D:/Thesis/files_output_dir/output_utilization/'

level1_road_history_workday_utility = pd.read_csv(output_dir + 'level1_road_history_workday_utilization_tune_AHP.csv')
level2_road_history_workday_utility = pd.read_csv(output_dir + 'level2_road_history_workday_utilization_tune_AHP.csv')
level3_road_history_workday_utility = pd.read_csv(output_dir + 'level3_road_history_workday_utilization_tune_AHP.csv')

level1_road_history_holiday_utility = pd.read_csv(output_dir + 'level1_road_history_holiday_utilization_tune_AHP.csv')
level2_road_history_holiday_utility = pd.read_csv(output_dir + 'level2_road_history_holiday_utilization_tune_AHP.csv')
level3_road_history_holiday_utility = pd.read_csv(output_dir + 'level3_road_history_holiday_utilization_tune_AHP.csv')

utility = level1_road_history_workday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['utilization'] > 0.7)].head(10)

utility = level1_road_history_workday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['time'] == 48) & (utility['utilization'] > 0.1) & (utility['road_id']== 2471)].head()

utility = level1_road_history_holiday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['time'] == 48) & (utility['utilization'] > 0.1) & (utility['road_id']== 2471)].head()

utility = level2_road_history_workday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['utilization'] > 0.6) & (utility['time'] == 48)][0:10]

utility = level2_road_history_workday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['time'] == 48) & (utility['utilization'] > 0.1) & (utility['road_id']== 6502)].head()

utility = level2_road_history_holiday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['time'] == 48) & (utility['utilization'] > 0.1) & (utility['road_id']== 6502)].head()

utility = level3_road_history_workday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[utility['utilization']>.7].head(10)

utility = level3_road_history_workday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['time'] == 48) & (utility['utilization'] > 0.1) & (utility['road_id']== 29933)].head()

utility = level3_road_history_holiday_utility[['road_id','time','avg_speed','volume','theta','vc', 'qc', 'Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm',
       'utilization' ]]
utility[(utility['time'] == 48) & (utility['utilization'] > 0.1) & (utility['road_id']== 29933)].head()
