==> 16. data sampling.ipynb <==
import pandas as pd

file_path ='D:/Thesis/files_output_dir/output_files/'
output_dir = 'D:/Thesis/files_output_dir/output_utilization/'

level1_road_history_workday_utility = pd.read_csv(output_dir + 'level1_road_history_workday_utilization.csv')

level1_road_history_workday_utility.columns

data_history = level1_road_history_workday_utility[['road_id', 'time','Uq', 'Uv', 'Uq_norm', 'Uv_norm', 'theta_norm', 'utilization']]
# data_history.head()

# Define the column to stratify by
stratify_column = 'time'

# Define the fraction of each stratum to sample
frac = 0.05  # 1% of each group

# Perform stratified sampling
stratified_sample = data_history.groupby(stratify_column, group_keys=False).apply(lambda x: x.sample(frac=frac))

# Display the stratified sample
stratified_sample.head()

stratified_sample.to_csv(output_dir + 'stratified_level1_road_history_workday_utility.csv',index=False)
